Added creationtime function

+++ Remove-TestAccounts.ps1
@@ -9,19 +9,22 @@ param (
 )
 $ErrorActionPreference = 'Stop'
 $user = Get-ADUser -filter "EmployeeID -eq '$EmployeeId'" -Properties *
-$user | Select-Object -Property SamAccountName, HomePage, departmentNumber
+if ($user){
 
-$gam = '.\bin\gam.exe'
-Write-Host ('{0},{1},Removing Gsuite' -f $MyInvocation.MyCommand.Name, $user.Homepage)
-& $gam delete user $user.HomePage
+ $user | Select-Object -Property SamAccountName, HomePage, departmentNumber
 
-Write-Host ('{0},{1},Removing AD' -f $MyInvocation.MyCommand.Name, $user.SamAccountName)
-$user | Remove-ADObject
+ $gam = '.\bin\gam.exe'
+ Write-Host ('{0},{1},Removing Gsuite' -f $MyInvocation.MyCommand.Name, $user.Homepage)
+ & $gam delete user $user.HomePage
 
-$lookupTable = Get-Content -Path .\json\site-lookup-table.json | ConvertFrom-Json
-$sc = $user.departmentNumber | Out-String
-$siteData = $lookupTable | Where-Object { [int]$_.SiteCode -eq [int]$sc }
-$msg = $MyInvocation.MyCommand.Name, $user.SamAccountName, $siteData.FileServer
-Write-Host ('{0},{1},{2},Removing Home Directory' -f $msg)
-. .\lib\Remove-StaffHomeDir.ps1
-Remove-StaffHomeDir -samid $user.SamAccountName -fileserver $siteData.FileServer -cred $FileServerCredential
+ Write-Host ('{0},{1},Removing AD' -f $MyInvocation.MyCommand.Name, $user.SamAccountName)
+ $user | Remove-ADObject
+
+ $lookupTable = Get-Content -Path .\json\site-lookup-table.json | ConvertFrom-Json
+ $sc = $user.departmentNumber | Out-String
+ $siteData = $lookupTable | Where-Object { [int]$_.SiteCode -eq [int]$sc }
+ $msg = $MyInvocation.MyCommand.Name, $user.SamAccountName, $siteData.FileServer
+ Write-Host ('{0},{1},{2},Removing Home Directory' -f $msg)
+ . .\lib\Remove-StaffHomeDir.ps1
+ Remove-StaffHomeDir -samid $user.SamAccountName -fileserver $siteData.FileServer -cred $FileServerCredential
+}
